+++ SnowflakeIDGenerator/Helpers/DateTimeHelper.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 
-namespace SnowflakeID
+namespace SnowflakeID.Helpers
 {
     internal static class DateTimeHelper
     {
         internal static ulong TimestampMillisFromEpoch(DateTime date, DateTime epoch)
         {
-            return ((ulong)date.Subtract(epoch).Ticks) / ((ulong)TimeSpan.TicksPerMillisecond);
+            return (ulong)date.Subtract(epoch).Ticks / TimeSpan.TicksPerMillisecond;
         }
     }
 }